chore: merge setup google sheet with sheet py
